Can add columns extracted and then deleted, and delete all rows then all columns

+++ src/io/HighsIO.cpp
@@ -59,13 +59,13 @@ void HighsLogMessage(HighsMessageType type, const char* format, ...) {
 
   if (logmsgcb == NULL) {
     fprintf(logfile, "%02d:%02d:%02d [%-7s] ", timeinfo->tm_hour, timeinfo->tm_min,
-            timeinfo->tm_sec, HighsMessageTypeTag[type]);
+            timeinfo->tm_sec, HighsMessageTypeTag[(int) type]);
     vfprintf(logfile, format, argptr);
     fprintf(logfile, "\n");
   } else {
 	int len;
 	len = snprintf(msgbuffer, sizeof(msgbuffer), "%02d:%02d:%02d [%-7s] ", timeinfo->tm_hour, timeinfo->tm_min,
-                   timeinfo->tm_sec, HighsMessageTypeTag[type]);
+                   timeinfo->tm_sec, HighsMessageTypeTag[(int) type]);
 	if (len < sizeof(msgbuffer))
 	  len += vsnprintf(msgbuffer + len, sizeof(msgbuffer) - len, format, argptr);
 	if (len < sizeof(msgbuffer)-1) {
